+++ js/src/index.ts
@@ -1,3 +1,7 @@
-export { default as useFlow } from './useFlow';
+export {
+  default as useFlow,
+  setErrorInterceptor,
+  UseFlowErrorInterceptor,
+} from './useFlow';
 export { default as flowAsState, State } from './flowAsState';
 export { Next, Next1, Next2, NextX } from './useFlow.types';