+++ kotlin/reakt-native-toolkit-ksp/src/main/kotlin/de/voize/reaktnativetoolkit/ksp/processor/TypescriptGenerator.kt
@@ -448,6 +448,12 @@ class TypescriptGenerator(
                                                                 unsubscribeFromToolkitUseFlow,
                                                             ).asCodeBlock()
                                                         )
+                                                        add(
+                                                            CodeBlock.of(
+                                                                "%S",
+                                                                "${exportedRNModuleName}.${functionDeclaration.simpleName.asString()}",
+                                                            )
+                                                        )
                                                         addAll(args)
                                                     }
                                                 )